+++ src/Pyro4/core.py
@@ -182,17 +182,17 @@ class Proxy(object):
     .. automethod:: _pyroHandshakeResponse
     .. autoattribute:: _pyroTimeout
     .. autoattribute:: _pyroHmacKey
+    .. attribute:: _pyroStuff
+
+        You can put custom data here without risking the problem of triggering remote attribute access. Pyro itself doesn't use this.
     """
     __pyroAttributes = frozenset(
         ["__getnewargs__", "__getnewargs_ex__", "__getinitargs__", "_pyroConnection", "_pyroUri",
          "_pyroOneway", "_pyroMethods", "_pyroAttrs", "_pyroTimeout", "_pyroSeq", "_pyroHmacKey",
-         "_pyroRawWireResponse",
+         "_pyroRawWireResponse", "_pyroStuff",
          "_Proxy__pyroHmacKey", "_Proxy__pyroTimeout", "_Proxy__pyroLock", "_Proxy__pyroConnLock"])
 
     def __init__(self, uri):
-        """
-        .. autoattribute:: _pyroTimeout
-        """
         if isinstance(uri, basestring):
             uri = URI(uri)
         elif not isinstance(uri, URI):
@@ -204,6 +204,7 @@ class Proxy(object):
         self._pyroOneway = set()  # oneway-methods of the remote object, gotten from meta-data
         self._pyroSeq = 0  # message sequence number
         self._pyroRawWireResponse = False  # internal switch to enable wire level responses
+        self._pyroStuff = None  # you can put custom data here without risking the problem of triggering remote attribute access. Not used by Pyro itself.
         self.__pyroHmacKey = None
         self.__pyroTimeout = Pyro4.config.COMMTIMEOUT
         self.__pyroLock = threadutil.Lock()
@@ -272,7 +273,7 @@ class Proxy(object):
                 self._pyroHmacKey = str(self._pyroHmacKey)
             encodedHmac = "b64:"+(base64.b64encode(self._pyroHmacKey).decode("ascii"))
         return self._pyroUri.asString(), tuple(self._pyroOneway), tuple(self._pyroMethods), tuple(self._pyroAttrs),\
-            self.__pyroTimeout, encodedHmac
+            self.__pyroTimeout, encodedHmac, self._pyroStuff
 
     def __setstate_from_dict__(self, state):
         uri = URI(state[0])
@@ -281,18 +282,19 @@ class Proxy(object):
         attrs = set(state[3])
         timeout = state[4]
         hmac_key = state[5]
+        stuff = state[6]
         if hmac_key:
             if hmac_key.startswith("b64:"):
                 hmac_key = base64.b64decode(hmac_key[4:].encode("ascii"))
             else:
                 raise errors.ProtocolError("hmac encoding error")
-        self.__setstate__((uri, oneway, methods, attrs, timeout, hmac_key))
+        self.__setstate__((uri, oneway, methods, attrs, timeout, hmac_key, stuff))
 
     def __getstate__(self):
-        return self._pyroUri, self._pyroOneway, self._pyroMethods, self._pyroAttrs, self.__pyroTimeout, self._pyroHmacKey  # skip the connection
+        return self._pyroUri, self._pyroOneway, self._pyroMethods, self._pyroAttrs, self.__pyroTimeout, self._pyroHmacKey, self._pyroStuff  # skip the connection
 
     def __setstate__(self, state):
-        self._pyroUri, self._pyroOneway, self._pyroMethods, self._pyroAttrs, self.__pyroTimeout, self._pyroHmacKey = state
+        self._pyroUri, self._pyroOneway, self._pyroMethods, self._pyroAttrs, self.__pyroTimeout, self._pyroHmacKey, self._pyroStuff = state
         self._pyroConnection = None
         self._pyroSeq = 0
         self._pyroRawWireResponse = False
@@ -308,6 +310,7 @@ class Proxy(object):
         p._pyroTimeout = self._pyroTimeout
         p._pyroHmacKey = self._pyroHmacKey
         p._pyroRawWireResponse = self._pyroRawWireResponse
+        p._pyroStuff = self._pyroStuff
         return p
 
     def __enter__(self):