+++ tests/PyroTests/test_serialize.py
@@ -178,6 +178,7 @@ class SerializeTests_pickle(unittest.TestCase):
         proxy._pyroOneway = set("ghi")
         proxy._pyroTimeout = 42
         proxy._pyroHmacKey = b"secret"
+        proxy._pyroStuff = [1, 2, 3]
         s, c = self.ser.serializeData(proxy)
         x = self.ser.deserializeData(s, c)
         self.assertIsInstance(x, Pyro4.core.Proxy)
@@ -187,6 +188,7 @@ class SerializeTests_pickle(unittest.TestCase):
         self.assertEqual(set("def"), x._pyroMethods)
         self.assertEqual(set("ghi"), x._pyroOneway)
         self.assertEqual(b"secret", x._pyroHmacKey)
+        self.assertEqual([1, 2, 3], x._pyroStuff)
         daemon = Pyro4.core.Daemon()
         s, c = self.ser.serializeData(daemon)
         x = self.ser.deserializeData(s, c)
@@ -210,12 +212,13 @@ class SerializeTests_pickle(unittest.TestCase):
         proxy._pyroOneway = set("ghi")
         proxy._pyroTimeout = 42
         proxy._pyroHmacKey = b"secret"
+        proxy._pyroStuff = [1, 2, 3]
         state = proxy.__getstate_for_dict__()
         if sys.platform == 'cli':
             b64_secret = "b64:"+base64.b64encode("secret").decode("utf-8")
         else:
             b64_secret = "b64:"+base64.b64encode(b"secret").decode("utf-8")
-        self.assertEqual(('PYRO:object@host:4444', tuple(set("ghi")), tuple(set("def")), tuple(set("abc")), 42, b64_secret), state)
+        self.assertEqual(('PYRO:object@host:4444', tuple(set("ghi")), tuple(set("def")), tuple(set("abc")), 42, b64_secret, [1, 2, 3]), state)
         proxy2 = Pyro4.core.Proxy("PYRONAME:xxx")
         proxy2.__setstate_from_dict__(state)
         self.assertEqual(proxy, proxy2)
@@ -225,6 +228,7 @@ class SerializeTests_pickle(unittest.TestCase):
         self.assertEqual(proxy._pyroOneway, proxy2._pyroOneway)
         self.assertEqual(proxy._pyroTimeout, proxy2._pyroTimeout)
         self.assertEqual(proxy._pyroHmacKey, proxy2._pyroHmacKey)
+        self.assertEqual(proxy._pyroStuff, proxy2._pyroStuff)
         daemon = Pyro4.core.Daemon()
         state = daemon.__getstate_for_dict__()
         self.assertEqual(tuple(), state)